+++ CHANGELOG.md
@@ -61,6 +61,7 @@ high state of flux, you're at risk of it changing without notice.
     - make `cons` pipeable (@gcanti)
   - `ReadonlyNonEmptyArray`
     - make `cons` pipeable (@gcanti)
+    - make `snoc` pipeable (@gcanti)
 
 # 2.9.1